Refactored language data representation
Changed how language data is represented. Instead of languages being
an Enum defined in constants, it is now a list of tuples defined in
settings. The reason for this change is it enables many django
functionalities related to language and translations. For example,
selecting a language to be used by a user on the site.

Added Tagalog as a custom language because it is not included as a
language in django.