Add language select in nav bar
Added a form for selecting language to the navigation bar. The reason we
want language selection is to provide translations for users to their
prefered language. Added to the navigation bar because that is visable
on every page of the site and we want for users to select language on
any page.