Use django-waffle feature flag for language select
Installed django-waffle, which allows for features to be enabled/disabled
based on a database entry. The purpose is to disable features while they
are being developted and then "turn them on" when they are ready.

Added a flag for language select in the navigation template. Because
selecting language currently has no effect, the select form will be
turned off.