Refactored choices from Enums to tuples
The reason for these changes is that there is a bug with using Enum value
as a choice "key" where the value is rejected by form validation for not
being a valid choice. To fix the issue, changed the Enum classes to normal
python classes and had the choices be a list of tuples of the class
atributes. The reason for using class attributes is that this is the
same way that all of the code is accessing the value of the choice, ie,
Languages.ENGLISH.

There was an issue of switching from using Enums in that migration files
were importing and using our Enum classes. To resolve this issue, I deleted
all migrations and created a new initial migration. Migrations no longer
rely on our code, so future changes to choices will not cause errors.