Add three fields and modified three others
The new fields are city, zip_code and is_ada_compliant. City and zip code
are requested options for filtering for events. Whether the event is
compliant with the ADA is also a requested data point.

The modified fields are contact_name, contact_email, contact_phone. They
have been modified to allow for their form entries to be blank. The
reason for this change is that there are some events that will not have
contact information, so we want for the contact info to be optional.