Change language filter from multiselect to dropdown
The multiselect checkboxes took up too much room due to the many
languages. To save on space, changed the selection to a dropdown menu.
The downside of doing this is that there can only be a single language
filtered at a time. The upside is that the form is much more condensed
with proper formating. This is a reasonable trade off.