Add copy event page
On the list of approved events, there is now a "Copy" button for each
event, which links to the copy page. The page contains a form with all
fields filled out from the existing event. All fields except for location
fields are editable. The reasoning for this is that copying an event
will be another occance of a very similar event in the same location.
On submit, the event is created as pending, just like a newly created
event.