Fix bug where all events filtered out
The bug was in the logic of what query parameters would be used to query
for events. The situation was when querying for events on a day in a
future month, the query would filter both for start datetime of the
future month as well as filter for events of the current month.