Add location to event overview
The event overview is the what is displayed for each of the events on
the home page. Changed the event title (part in bold) to be the event
title instead of start - end times. Added the address location as a
new line. It is important to include the event location because it is
one of the main things people care about when considering going to an
event.